Refactor: 화살 asset
