+++ frontend/src/components/Sidebar/OptionItem/index.tsx
@@ -2,7 +2,7 @@ import { useCookies } from 'react-cookie';
 import styled from 'styled-components';
 
 import { ReactComponent as ArrowDown } from '../../../assets/icons/arrowDown.svg';
-import ArrowRight from '../../../assets/icons/arrowRight';
+import ArrowRight from '../../../assets/icons/arrowRight.svg?component';
 import { ReactComponent as ArrowUp } from '../../../assets/icons/arrowUp.svg';
 import PALETTE from '../../../constants/palette';
 import { Children, OptionDecoration } from '../../../types';
