+++ frontend/src/types/assets.d.ts
@@ -1,3 +1,4 @@
 declare module '*.png';
 declare module '*.svg';
+declare module '*.svg?component';
 declare module '*.gif';
