+++ frontend/vite.config.ts
@@ -4,5 +4,5 @@ import svgr from 'vite-plugin-svgr';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(), svgr()],
+  plugins: [svgr(), react()],
 });
